Break out top row into its own component
