+++ src/components/paint-editor/paint-editor.css
@@ -28,17 +28,9 @@
     margin-top: calc(2 * $grid-unit);
 }
 
-.costume-input {
-    width: 8rem;
-}
-
 .mod-dashed-border {
     border-right: 1px dashed $ui-pane-border;
     padding-right: calc(2 * $grid-unit);
-}
-
-.mod-unselect {
-    user-select: none;
 }
 
 .mod-labeled-icon-height {
@@ -78,43 +70,6 @@ $border-radius: 0.25rem;
     width: 1.25rem;
     height: 1.25rem;
     vertical-align: middle;
-}
-
-.mod-context-menu {
-    display: flex;
-    flex-direction: column;
-}
-
-.mod-top-divider {
-    border-top: 1px solid $ui-pane-border;
-}
-
-.mod-menu-item {
-    display: flex;
-    margin: 0 -$grid-unit;
-    min-width: 6.25rem;
-    padding: calc(3 * $grid-unit);
-    white-space: nowrap;
-    cursor: pointer;
-    transition: 0.1s ease;
-    align-items: center;
-    font-family: "Helvetica Neue", Helvetica, sans-serif;
-}
-
-.mod-disabled {
-    cursor: auto;
-}
-
-.mod-menu-item:hover {
-    background: $motion-transparent;
-}
-
-.mod-disabled:hover {
-    background-color: transparent;
-}
-
-.menu-item-icon {
-    margin-right: calc(2 * $grid-unit);
 }
 
 .mod-mode-tools {
@@ -173,11 +128,6 @@ $border-radius: 0.25rem;
     display: flex;
     margin-top: .25rem;
     justify-content: space-between;
-}
-
-.bitmap-tooltip {
-    margin-left: $grid-unit;
-    max-width: 10rem;
 }
 
 .bitmap-button {
